restructuring controllers & setting productPrice on orderDetail
